# apps/server/docs/swagger.yaml
definitions:
  dto.ExperienceDto:
    properties:
      company:
        type: string
      position:
        type: string
      responsibilities:
        type: string
      timeFrame:
        type: string
    type: object
  dto.ResponseDto:
    properties:
      message:
        type: string
    type: object
  dto.TechnologyDto:
    properties:
      description:
        type: string
      experience:
        type: string
      logoClass:
        type: string
      name:
        type: string
    type: object
  dto.TelemetryDto:
    properties:
      countryIso:
        type: string
      countryName:
        type: string
      timestamp:
        type: string
    type: object
  dto.UserDto:
    properties:
      password:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/authentication:
    post:
      consumes:
      - application/json
      description: Authenticate user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserDto'
      produces:
      - application/json
      responses:
        "201":
          description: User with token
          schema:
            $ref: '#/definitions/dto.UserDto'
        "500":
          description: Internal Server Error
      summary: Authenticate user
      tags:
      - user
  /api/experience:
    delete:
      consumes:
      - application/json
      description: Deletes experience
      parameters:
      - description: Experience id
        in: path
        name: id
        required: true
        type: number
      - description: BasicAuth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/dto.ResponseDto'
        "500":
          description: Internal Server Error
      summary: Delete experiences
      tags:
      - experience
    get:
      consumes:
      - application/json
      description: Gets experiences
      produces:
      - application/json
      responses:
        "200":
          description: Experiences
          schema:
            items:
              $ref: '#/definitions/dto.ExperienceDto'
            type: array
        "500":
          description: Internal Server Error
      summary: Get experiences
      tags:
      - experience
    post:
      consumes:
      - application/json
      description: Creates experience
      parameters:
      - description: Experience
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserDto'
      - description: BasicAuth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created experience
          schema:
            $ref: '#/definitions/dto.ExperienceDto'
        "500":
          description: Internal Server Error
      summary: Create experience
      tags:
      - experience
    put:
      consumes:
      - application/json
      description: Updates experience
      parameters:
      - description: Experience
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.ExperienceDto'
      - description: BasicAuth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated experience
          schema:
            $ref: '#/definitions/dto.ExperienceDto'
        "500":
          description: Internal Server Error
      summary: Update experiences
      tags:
      - experience
  /api/technology:
    delete:
      consumes:
      - application/json
      description: Deletes technology
      parameters:
      - description: Technology id
        in: path
        name: id
        required: true
        type: number
      - description: BasicAuth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/dto.ResponseDto'
        "500":
          description: Internal Server Error
      summary: Delete technology
      tags:
      - technology
    get:
      consumes:
      - application/json
      description: Gets technologies
      produces:
      - application/json
      responses:
        "200":
          description: Technologies
          schema:
            items:
              $ref: '#/definitions/dto.TechnologyDto'
            type: array
        "500":
          description: Internal Server Error
      summary: Get technologies
      tags:
      - technology
    post:
      consumes:
      - application/json
      description: Creates technologies
      parameters:
      - description: Technology
        in: body
        name: technology
        required: true
        schema:
          $ref: '#/definitions/dto.TechnologyDto'
      - description: BasicAuth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created technology
          schema:
            $ref: '#/definitions/dto.TechnologyDto'
        "500":
          description: Internal Server Error
      summary: Create technology
      tags:
      - technology
    put:
      consumes:
      - application/json
      description: Updates technologies
      parameters:
      - description: Technology
        in: body
        name: technology
        required: true
        schema:
          $ref: '#/definitions/dto.TechnologyDto'
      - description: BasicAuth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated technology
          schema:
            $ref: '#/definitions/dto.TechnologyDto'
        "500":
          description: Internal Server Error
      summary: Update technology
      tags:
      - technology
  /api/telemetry:
    get:
      consumes:
      - application/json
      description: Gets telemetries
      produces:
      - application/json
      responses:
        "200":
          description: Telemetries
          schema:
            items:
              $ref: '#/definitions/dto.TelemetryDto'
            type: array
        "500":
          description: Internal Server Error
      summary: Get telemetries
      tags:
      - telemetry
    post:
      consumes:
      - application/json
      description: Creates telemetry
      parameters:
      - description: Telemetry
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.TelemetryDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created telemetry
          schema:
            $ref: '#/definitions/dto.ExperienceDto'
        "500":
          description: Internal Server Error
      summary: Create telemetry
      tags:
      - telemetry
swagger: "2.0"